Add com(4) terminal and console support.
Implement com(4) interrupt support for receiving data.

Add termios terminal support to the com(4) driver with tcsetattr(2)
support for configuring the hardware serial settings.

Add kernel(7) --console option which can be used to initialize the
com(4) driver early for the kernel console with the specified serial
settings and window size.

Add kernel(7) --text option for overriding the TERM environment variable.

Add kernel(7) --disable-logo option for disabling writing the the
operating system logo to the console on boot.

Add advanced bootloader option to select booting to a serial console on
com1. Add bootloader variables and hooks to customize this behavior.

Add tix-iso-bootconfig(8) --console, --grub-serial, --serial, and
--serial-console options as a convenince to opt into booting to the
serial console by default.

Add tix-iso-bootconfig(8) --kernel-options option for forwarding
additional options to the kernel.

The logterminal driver used for tty1 currently assumes that it controls
the kernel log, which is not true when com(4) is used as the console,
therefore the tty1 driver is currently disabled if a serial console is
selected, as a temporary workaround.