Add init(8) --prefix and --static-prefix options.
This change refactors init's internals to use fine grained directory paths.

The new --static-prefix option provides sysmerge(8) --booting with a
well defined environment that uses the local data from the root
directory with the new static data of the /sysmerge directory. This way
the upgrade honors the local configuration while ensuring no programs
with the wrong ABI are executed during the upgrade process. The goal is to
reduce the room for bugs by avoiding running the upgrade inside a weird
ill-defined environment where standard assumptions might not be true.