Add fatfs(8).
This change implements the FAT filesystem as a user-space filesystem deamon.

The FAT filesystem is required by EFI as the EFI System Partition (ESP)
needed in order to boot on EFI systems.

fatfs(8) is case sensitive and does not implement case sensitivity. The
long names are used if available, and the short file name space is not
available unless files do not have a short name. It is possible to create
filesystem contents that do not interoperate with other drivers.

fatfs(8) only uses the UTC timezone, whereas FAT officially is local time.

fatfs(8) uses the first cluster of files as the inode number, since it is
permanent even when moved across directories. If any zero size files are
seen with no first cluster, then it is allocated a first cluster, so it
has a permanent inode number. This behavior is not official FAT behavior
and fsck.fat(8) will undo it as filesystem damage.

fatfs(8) uses FAT entry 1 to detect if it wasn't properly unmounted. This
behavior is official on FAT16/FAT32, but fatfs(8) implements it as an
unofficial extension on FAT12 as well. It's important to ensure that
a damaged filesystem is properly fixed with fsck.fat(8).

The volume label in the root directory is not used due to the complexity
of reading it inside libmount. The volume label in the boot sector is used
instead, and can be set with fatlabel(8), but Windows does not update the
label in the boot sector.

fsync(2) is not properly implemented, as fatfs(8) has the same
architectural design problems as extfs(8), which will be fixed together in
the future.

libmount can now recognize regular FAT filesystems and the ESP.