Fix tix-port(8) DEVELOPMENT putting too many details in .version.
The hashes of the .port and patches might change a lot during development
and if they ever get out of sync, then tix-port(8) would annoyingly give a
fatal error that it wanted to delete the source code but can't because the
port is in development. This change makes the workflow much more useful.