Fix unix(4) handling of sun_path without nul or too long.
A sun_path without a nul byte could panic the kernel because the
path_offset - addrsize expression underflowed, leading to a buffer overrun
leading to undefined behavior such as a crash.

Limit the extra-long-unix-path extension to 4096 bytes for safety. Avoid
using sun_path to access the path, since it's limited in size, and reading
it beyond bounds constitutes undefined behavior.