Add getaddrinfo(3) AI_DNS and AI_DNSRAW flags.
The new struct addrinfo fields ai_ttl, ai_dnsclass, and ai_dnstype provide
access to all DNS records with full information. Previously important
records such as MX and TXT could not be accessed through getaddrinfo(3),
and there was no information on how long the answer would be valid.

If AI_DNS or AI_DNSRAW is set in hints, then the specific ai_dnsclass and
ai_dnstype questions (following the ai_next chain) are sent to the server.
These are the standard DNS QCLASS and QTYPE values.

AI_DNS is set in the output addrinfo, if the ai_ttl, ai_dnsclass, and
ai_dnstype fields have been set and exists. The fields should not be
accessed if AI_DNS is not set. If any answers cannot be converted to a
struct sockaddr, and AI_DNS is set in hints, then the raw answers are
included in the output chain with AI_DNS set in ai_flags, and ai_family set
to AF_UNSPEC. If AI_DNSRAW is set in hints, then answers are not converted
to struct sockaddr, CNAME chains are not followed, and the raw DNS response
is returned.