Fix setpgid(2) semantics for standards compliance.
It's allowed to move a group leader out of its own group.

If a child has run execve(2), then changing its process group from the
parent must fail with EACCES.

Zombie processes are not removed from the group/session/init until they
are awaited.

Fail with ESRCH instead of EPERM if the process is not the current process
or a direct child.