Fix tcsetpgrp(2) and tcgetpgrp(2) semantics for standards compliance.
PIX_MAX rather than -1 is used when a terminal does not have a foreground
group, as POSIX requires a value larger than 1 that is not a valid pgid.

The syscalls now enforce the processes and groups are in the right session.

tcgetsid(2) is fixed so it fails ENOTTY if no controlling terminal is set.