Add <limits.h>.
gcc needs to be rebuilt to detect the existence of <limits.h>. gcc will wrap
the <limits.h> header unconditionally, and include the system <limits.h>
only if it detected the system header at gcc configure time.

It's safe for the system <limits.h> to define everything it's supposed to,
since gcc will undef anything it wants to provide itself. All the constants
outside of the C programming language are put in the <sortix/limits.h>
header that is indirectly included.

For bootstrap compatibility using previous Sortix compilers, all source
files that need POSIX extensions to <limits.h> will temporarily also include
the <sortix/limits.h> header if required. This support is required to
upgrade from source using 1.0 or any version in-between.