Run global destructors in exit(3) and cut on_exit and pthread dependencies.
exit(3) now invokes the fini function after the on_exit(3) destructors and
before the final FILE flush. This change makes exit(3) semantically correct.

Weak aliases are now uses to only pull in pthread locking if pthread_create
is in the executable, and likewise to only invoke the on_exit(3) destructors
if on_exit(3) is in the executable. These changes reduces the minimal static
link size.