Support threads in libstdc++.
Fix the libstdc++ configure checks with g++ failing because it's linking
with libstdc++ before it's already built.

Work around libstdc++ assuming libgcc has just been built with threading
support, so the configure script can cheat and steal libgcc's homework,
but that doesn't work when libstdc++ is built standalone.

Expose the new POSIX.1-2024 pthread functions in C++ mode, because
libstdc++ can use them and uses them even in programs not compiled
with POSIX.1-2024 features. I can't quite think of a cleaner solution
with the current design of the C++ standard library.