Improve kernelinfo(2)'s formatting.
I've started using .Qq for /all/ strings, I wasn't consistent about it
in the original patch for some reason.

The list was sized based on the width of "options", which seems to be
the sweet spot when typesetting for 80 characters of width.

I've also moved the builddate/time examples after the mention of the
macro, it was a bit ungrammatical before.  I could've also just added
commas after the examples, but I like this more.

I still don't like the duplication in RETURN VALUES but I guess that's
fine.