Remove /home and /mnt from the system manifest.
The /home, /mnt, and /src directories are no longer part of the system
package and are rather created by sysinstall(8). Operating system upgrades
will temporarily remove the directories if they are empty. No data will be
deleted under any circumstances. A post-upgrade hook will recreate the /home
and /mnt directories if they were empty and ownership is transferred from
the operating system to the system administrator.