Build and install all system files with exact permissions.
This change removes the implict dependency on the umask being 0022 to
produce a system with the correct permissions. install(1) cannot portably
copy directory trees, so fix up the permissions after cp with chmod loops
on every file.

Luckily the 1.0 release utilities actually have all the required options,
although the install(1) options are ignored, which means this change can
be bootstrapped without any issues.