Add execveat(2) and fexecve(2).
fexecve uses argv[0] when invoking the #! interpreter since there is
currently no other way to pass a fd to a program.

Fall back on the old execve syscall if execveat is unable for bootstrap
issues due to too many ports with native sysroot issues.

Allow reading a file with O_EXEC only.

This is a compatible ABI change. The old syscall is retained for temporary
ABI compatibility with the 1.0 stable release, and will be removed after the
next stable release.