Add getgroups(2) and setgroups(2) system calls.
Raise NGROUPS_MAX to 65536 to match the Linux limit, but still have a limit
to start working against unbounded allocations in the kernel.

The non-standard setgroups(2) is put in <unistd.h> to match the BSD
behavior of putting it alongside getgroups(2), whereas glibc puts it in
<grp.h> for some reason which feels unexpected.

The groups are not actually used for anything at this time, and there is no
actual security enforcement, which is well known. The actual security will
be implemented later, but it is useful to have the system calls ahead of the
next stable release.

Add psctl(2) PSCTL_GROUPS so other processes are able to inspect the
supplementary groups of processes. This feature could be implemented later
in e.g. ps(1).

Remove workarounds from ports that use getgroups(2) and setgroups(2).
The related initgroups(3) is not implemented at this time, as it is not a
system call.

This is a compatible ABI change.