Fix tix-upgrade(7) accidentally glob expanding all package dependencies.
The 'all' package depends on '*' as a special dependency that means all
packages, but the shell for loop accidentally did not escape the character
and the usual globbing happened. The end result was that tix-upgrade(7) did
not work correctly if executed in a non-empty directory if the all package
was installed, as it would otherwise start installing every filename in the
current directory as if they were packages.